requirements.txt
absl-py==1.4.0
aio-pika==8.2.3
aiofiles==24.1.0
aiogram==2.15
aiohttp==3.9.5
aiormq==6.4.2
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.28.1
anyio==4.7.0
APScheduler==3.9.1.post1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
azure-cognitiveservices-speech==1.41.1
babel==2.16.0
bidict==0.23.1
blinker==1.9.0
boto3==1.35.76
botocore==1.35.76
CacheControl==0.12.14
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==2.2.1
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confluent-kafka==2.6.1
cryptography==44.0.0
cycler==0.12.1
dask==2022.10.2
Deprecated==1.2.15
distro==1.9.0
dnspython==2.3.0
docopt==0.6.2
elevenlabs==1.13.3
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.6
fbmessenger==6.0.0
filelock==3.16.1
fire==0.7.0
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.55.2
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
gast==0.4.0
google-auth==2.36.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.68.1
gTTS==2.5.4
h11==0.14.0
h5py==3.12.1
html5tagger==1.3.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.26.5
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
janus==1.1.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
jiter==0.8.0
jmespath==1.0.1
joblib==1.2.0
jsonpickle==3.0.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==2.12.0
kiwisolver==1.4.7
libclang==18.1.1
locket==1.0.0
loguru==0.7.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.5.3
mattermostwrapper==2.2
mdurl==0.1.2
miniaudio==1.61
ml_dtypes==0.5.0
moto==5.0.22
msgpack==1.1.0
multidict==6.1.0
neo4j==5.27.0
networkx==2.6.3
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.57.0
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opt_einsum==3.4.0
packaging==20.9
pamqp==3.2.1
partd==1.4.2
phonenumbers==8.13.51
pillow==11.0.0
pluggy==1.5.0
portalocker==2.10.1
prompt-toolkit==3.0.28
propcache==0.2.1
protobuf==4.25.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.13
pycparser==2.22
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydot==1.4.2
pydub==0.25.1
Pygments==2.18.0
pyht==0.0.28
PyJWT==2.10.1
pykwalify==1.8.0
pymongo==4.3.3
pyparsing==3.2.0
pyrsistent==0.20.0
python-crfsuite==0.9.11
python-dateutil==2.8.2
python-dotenv==0.21.1
python-engineio==4.10.1
python-multipart==0.0.19
python-socketio==5.11.4
pytz==2022.7.1
PyYAML==6.0.2
questionary==1.10.0
randomname==0.1.5
rasa==3.6.20
rasa-sdk==3.6.2
redis==5.2.1
referencing==0.35.1
regex==2022.10.31
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
responses==0.25.3
rich==13.9.4
rich-toolkit==0.12.0
rocketchat-API==1.30.0
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.12
s3transfer==0.10.4
sanic==24.6.0
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==23.12.0
scikit-learn==1.1.3
scipy==1.10.1
sentry-sdk==2.19.2
shellingham==1.5.4
signalwire==2.0.4
simple-websocket==1.1.0
six==1.17.0
sklearn-crfsuite==0.3.6
slack_sdk==3.33.5
sniffio==1.3.1
sounddevice==0.4.7
SQLAlchemy==1.4.54
starlette==0.37.2
structlog==23.3.0
structlog-sentry==2.1.0
tabulate==0.9.0
tarsafe==0.0.4
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-text==2.12.0
termcolor==2.5.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.21.0
toolz==1.0.0
tqdm==4.67.1
tracerite==1.1.1
twilio==6.43.0
typer==0.15.1
typing-utils==0.1.0
typing_extensions==4.12.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.21.0
vocode==0.1.113
watchfiles==1.0.0
wcwidth==0.2.13
webexteamssdk==1.6.1
websockets==14.1
Werkzeug==3.1.3
whisper==1.1.10
wrapt==1.14.1
wsproto==1.2.0
xmltodict==0.14.2
yarl==1.18.3
zipp==3.21.0
